+++ R/layer-brain.R
@@ -172,7 +172,7 @@ LayerBrain <- ggproto(
     }
 
     if (needs_mapping("fill")) {
-      self$computed_mapping$fill <- as.name("region")
+      self$computed_mapping$fill <- as.name("label")
     }
 
     self$computed_mapping$label <- as.name("label")